restructured naming for clarity
